Move EventLog components to own folder
