+++ client/src/views/event-log/EventLogNewAssignment.tsx
