+++ client/src/views/event-log/EventLogNoAssignment.tsx
